requirements.txt
bcrypt==4.2.0
pillow==11.0.0
setuptools==75.3.0